requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
beautifulsoup4
python-dotenv
pytest
httpx
pydantic-settings
pydantic[email]
